Update SimpleHartMgr.py
This guy connects to a HART manager, but it doesn't do anything.  I need help with importing the DEFAULT connection parameters correctly.  It works when you type in a real ip address and type in 4445 but I had to hardcode the default user and password because I couldn't import them correctly